// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/attention-signal-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AttentionSignalHttpResponse {
    clientDisplayName?: string | null;
    clientUserId?: string;
    detectedAtUtc?: string;
    id?: string;
    reason?: string;
    severity?: string;
    snoozedUntilUtc?: string | null;
    type?: string;
}