// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/bulk-recommendation-result-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BulkRecommendationRecipientResultHttpResponse } from './bulk-recommendation-recipient-result-http-response';

export interface BulkRecommendationResultHttpResponse {
    idempotencyKey?: string;
    recipients?: Array<BulkRecommendationRecipientResultHttpResponse>;
}