// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/client-summary-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DietologistPermissionsHttpResponse } from './dietologist-permissions-http-response';

export interface ClientSummaryHttpResponse {
    acceptedAtUtc?: string;
    activityLevel?: string | null;
    birthDate?: string | null;
    email?: string | null;
    firstName?: string | null;
    gender?: string | null;
    heightCm?: number | null;
    lastName?: string | null;
    permissions?: DietologistPermissionsHttpResponse;
    profileImage?: string | null;
    userId?: string;
}