// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/model/dietologist-relationship-http-response.ts
/**
 * FoodDiary User API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DietologistPermissionsHttpResponse } from './dietologist-permissions-http-response';

export interface DietologistRelationshipHttpResponse {
    acceptedAtUtc?: string | null;
    createdAtUtc?: string;
    dietologistUserId?: string | null;
    email?: string | null;
    expiresAtUtc?: string;
    firstName?: string | null;
    invitationId?: string;
    lastName?: string | null;
    permissions?: DietologistPermissionsHttpResponse;
    status?: string;
}